Clean up metrics and move them to separate files
